moved matrix prior model to its own module
